% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_ranges}
\alias{all_ranges}
\title{all ranges}
\format{
Simple feature collection with 6 features and 1 field
geometry type:  GEOMETRY
dimension:      XY
bbox:           xmin: -78.6167 ymin: -30.27966 xmax: -36.88333 ymax: 0.3333
epsg (SRID):    4326
proj4string:    +proj=longlat +datum=WGS84 +no_defs
\describe{
\item{\code{tax_accepted_name}}{the scientific species name}
\item{\code{geometry}}{the range polygon of the species}
}
}
\usage{
all_ranges
}
\description{
Polygons of the geographic distribution for 3295 bromeliad species as derived from distribution modelling.
For species with more than 9 occurrence records available modelled using climate SDMs,
for species with 3-8 records available modelled by convex hulls, and
for species with less than 3 records modelled by a geographic buffer with 50 kilometer radius
}
\keyword{datasets}
