% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{georef_additions}
\alias{georef_additions}
\title{georef_additions}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1169 rows and 13 columns.}
\usage{
georef_additions
}
\description{
A data.frame of occurrence records derived from georeferencing type specimen from the literature.
}
\details{
\describe{
\item{\code{accepted_name_species}}{}
\item{\code{scientificname}}{}
\item{\code{family}}{}
\item{\code{genus}}{}
\item{\code{species}}{}
\item{\code{taxonrank}}{}
\item{\code{decimallongitude}}{}
\item{\code{decimallatitude}}{}
\item{\code{coordinate_accuracy}}{}
\item{\code{flag}}{}
\item{\code{remarks}}{}
\item{\code{basisofrecords}}{}
\item{\code{source}}{}
}
}
\keyword{datasets}
