% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range.R
\name{get_range}
\alias{get_range}
\title{Select Bromeliaceae Species Ranges by Taxonomy and Traits}
\usage{
get_range(
  scientific = NULL,
  canonical = NULL,
  genus = NULL,
  subfamily = NULL,
  life_form = NULL,
  assessment = NULL,
  range_size = NULL,
  type = "polygon"
)
}
\arguments{
\item{scientific}{a character vector of full scientific names including authorities, of the species of interest}

\item{canonical}{a character vector of canonical names, of the species of interest.}

\item{genus}{a character vector of genera names to select.}

\item{subfamily}{a character vector of subfamily names to select.}

\item{life_form}{a character vector of life forms to select.}

\item{assessment}{a character vector of conservation assessment to select.}

\item{range_size}{a vector of two numericals with the minimum and maximum range size (in square kilometres), to select.}

\item{type}{a cahracter defining the output format, see details}
}
\value{
Depending on the \dQuote{type} argument. If \dQuote{binary} a presence/absence raster based on the modelled habitat suitability,
 at 100x100 km resolution in Behrmann projection,
 if \dQuote{suitability} the habitat suitability as predicted by an ensemble model in Behrmann projection, and
 if {polygon} a simple feature object in lat/lon projection.
.
}
\description{
Get the geographic range for all species selected via the arguments.
The type of range estimate depends on the \dQuote{type} argument.
}
\details{
Modelled ranges are available for 542 species,
range polygons for 2395 species. For species with model distribution, the range polygons are based on the models, otherwise
on a convex hull around available occurrence records, or a 50 km buffer for species with less than 3 occurrence records available.
See Zizka et al 2019 for methods.
}
\examples{
get_range(scientific = "Aechmea mexicana Baker")
get_range(scientific = "Aechmea mexicana Baker", type = "binary")
get_range(scientific = "Aechmea mexicana Baker", type = "suitability")
get_range(canonical = "Aechmea mexicana")
get_range(genus = "Aechmea")
get_range(genus = "Aechmea", type = "binary")
get_range(genus = c("Aechmea", "Zizkaea"))
get_range(subfamily = "Pitcairnioideae")
get_range(life_form = "epiphyte")
get_range(assessment = c("CR", "VU"))
get_range(range_size = c(1000, 10000))

}
