% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_richness.R
\name{map_richness}
\alias{map_richness}
\title{Mapping Bromeliad Species Richness}
\usage{
map_richness(x, title = NULL)
}
\arguments{
\item{x}{an object of the class \dQuote{bromeli} generated by get_range.}

\item{title}{An optional title to be added to the plot.}
}
\value{
aa richness map of the selected taxa across the Neotropics.
.
}
\description{
Produces a richness map in a 100x100 km grid for the selected bromeliaceae species.
}
\examples{
map_richness(x = get_range(genus = "Aechmea", type = "binary"))

}
