% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_species.R
\name{map_species}
\alias{map_species}
\title{Mapping species distributions}
\usage{
map_species(scientific = NULL, canonical = NULL, genus = NULL,
  subfamily = NULL, life_form = NULL, assessment = NULL,
  range_size = NULL, model = "suitability", polygon = TRUE,
  iucn = TRUE)
}
\arguments{
\item{scientific}{a character vector of full scientific names including authorities, of the species of interest}

\item{canonical}{a character vector of canonical names, of the species of interest.}

\item{genus}{a character vector of genera names to select.}

\item{subfamily}{a character vector of subfamily names to select.}

\item{life_form}{a character vector of life forms to select.}

\item{assessment}{a character vector of conservation assessment to select.}

\item{range_size}{a vector of two numericals with the minimum and maximum range size (in square kilometres), to select.}

\item{model}{character string. If \dQuote{binary} presence absence per 100 km grid cell is plotted,
if \dQuote{suitability} the modelled habitat suitability is plotted.}

\item{polygon}{logical. If TRUE, an outline of the range polygon is plotted.}

\item{iucn}{logical. If TRUE a legend with the conservation assessment is added to the plot.}
}
\value{
a plot for each selected species
.
}
\description{
Produces distribution maps for all species in the selection
}
\details{
Modelled ranges are available for 542 species,
range polygons for 2395 species. For species with model distribution, the range polygons are based on the models, otherwise
on a convex hull around available occurrence records, or a 50 km buffer for species with less than 3 occurrence records available.
See Zizka et al 2019 for methods.
}
\examples{
map_species(canonical = "Aechmea mexicana")

}
