% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_binary}
\alias{model_binary}
\title{model_binary}
\format{
An object of class \code{list} of length 581.
}
\usage{
model_binary
}
\description{
A list of raster layers, including the projected binary distributions for 541 species as defined by ensemble modelling.
}
\keyword{datasets}
