% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomy_traits}
\alias{taxonomy_traits}
\title{taxonomy_traits}
\format{
A data frame with 17 variables:
\describe{
\item{\code{tax_accepted_name}}{the accepted species name}
\item{\code{tax_genus}}{the genus}
\item{\code{tax_tribus}}{the tribus}
\item{\code{tax_subfamily}}{the subfamily}
\item{\code{geo_range}}{the range size of the species, based on the modelled distribution}
\item{\code{conv_conr_eoo}}{the extent of occurrence}
\item{\code{conv_conr_aoo}}{the area of occupancy}
\item{\code{conv_conr_pop}}{the number of independent populations deived during automated assessment}
\item{\code{conv_conr_code}}{The detailled code of the conservation status based on automated assessment}
\item{\code{conv_conr_stat}}{The conservation status of the species based on automated conservation assessment}
\item{\code{tr_growth_form}}{the growth form of the species from the world checklist of selected plant families}
\item{\code{tax_canonical}}{the canonical species name}
}
}
\usage{
taxonomy_traits
}
\description{
A datset of species known to be phylogenetically derived woody
}
\keyword{datasets}
